Move common code to parent header
